Mule-ManageService--Python-Version/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.6.0
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.4